+++ sdks/python/apache_beam/options/value_provider_test.py
@@ -233,11 +233,11 @@ class ValueProviderTests(unittest.TestCase):
       mock_fn()
       return x
 
-    vp = NestedValueProvider(StaticValueProvider(int, 1), lambda x: x + 1)
+    vp = NestedValueProvider(StaticValueProvider(int, 1), translator)
 
-    vp.get_value()
+    vp.get()
     self.assertEqual(mock_fn.call_count, 1)
-    vp.get_value()
+    vp.get()
     self.assertEqual(mock_fn.call_count, 1)
 
   def test_nested_value_provider_wrap_runtime(self):